Update the transtion() section.
These changes reflect the latest on how to give the caller the most
flexibility in choosing an EME pipleine.

For example: say that your key system has both a hw secure and sw
pipeline. Transitioning between these pipelines may be unsupported (due
to having the keys tightly coupled to a given pipeline).

In the previous proposal, we choose the pipeline based on the first
decodingInfo() configuration and *stuck to that choice*. The concern is
that this was too inflexible given that the caller had not yet had an
opportunuity to provide other configurations (via transition()). If we
could see into the future transition() calls, we might have selected a
pipeline that supported the combined configurations instead of just
optimizing for the first.

This update gives more flexibility by chaing the semantics of
transition(). Each transtion() call now serves as a signal that the
caller desires to use this additioanl configuration and implementers
should re-run the a pipeline selection algo in an attempt to support
the combined configs.

This update also changes the meaning of the the MediaCapabilityInfo
booleans that stem from a transition() call. Rather than describing only
the support/performance of the single config provided to transition(),
these boolenas now describe the lowest denominator of
support/performance across all configs provided to decodingInfo() and
chained transition() calls. This allows us to surface when performance
(e.g. powerEfficient) is altered as a result of choosing a new pipeline
to support the transition() configuration.