Fix automatic publication settings
This fixes the settings to automate the publication of the spec. Based on
discussions with @deniak and experience with automating the publication of
Picture in Picture:
https://github.com/w3c/picture-in-picture/pull/200

Updates are:
- Add a "Previous Version" link
- Set the spec level to `None` because the spec has not been published as a
leveled spec (we may change that later on if there ever comes a time when the
group decides to publish a level 2), and `Level: 1` tells the publication system
to publish under `media-capabilities-1`, which does not exist.
- Force the status to WD during publication, otherwise the publication system
sees an ED and refuses to publish the document.