Adjust ref to secure context and shrink hardcoded xref links
The Secure Contexts spec no longer has an algorithm named "Is the environment
settings object settings a secure context?". That part got merged into HTML.
This update adjusts the text accordingly, fixing the broken link.

It also drops custom definitions of all terms that are now part of Bikeshed's
autolinking database. Remaining ones are needed because underlying specs are
not yet in the database.